Added task complete feature
